+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add parameter `--print-topics` to `describe consumer-groups` to control if topics shall be printed
 - Add alias `list` for `get` command (e.g. `list topics`)
 - kafkactl is now installable with `homebrew`
+- a docker image is now generated with every release
 
 ## 1.5.0 - 2020-01-24